derivation/src/tests/mod.rs
use super::{serialize_derivation, Derivation};
use std::fs::File;
use std::io::Read;
use std::path::Path;
use test_generator::test_resources;

fn read_file(path: &str) -> String {
    let path = Path::new(path);
    let mut file = File::open(path).unwrap();
    let mut data = String::new();

    file.read_to_string(&mut data).unwrap();

    return data;
}

fn assert_derivation_ok(path_to_drv_file: &str) {
    let data = read_file(&format!("{}.json", path_to_drv_file));
    let derivation: Derivation = serde_json::from_str(&data).expect("JSON was not well-formatted");

    let mut serialized_derivation = String::new();
    serialize_derivation(derivation, &mut serialized_derivation).unwrap();

    let expected = read_file(path_to_drv_file);

    assert_eq!(expected, serialized_derivation);
}

#[test_resources("src/tests/derivation_tests/*.drv")]
fn derivation_files_ok(path: &str) {
    assert_derivation_ok(path);
}
